+++ src/Sulu/Bundle/SnippetBundle/Controller/SnippetController.php
@@ -13,20 +13,24 @@ namespace Sulu\Bundle\SnippetBundle\Controller;
 
 use FOS\RestBundle\Controller\Annotations\Get;
 use FOS\RestBundle\Controller\Annotations\Post;
+use FOS\RestBundle\Routing\ClassResourceInterface;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\View\ViewHandler;
+use JMS\Serializer\SerializationContext;
 use PHPCR\NodeInterface;
 use Sulu\Bundle\SnippetBundle\Snippet\DefaultSnippetManagerInterface;
 use Sulu\Bundle\SnippetBundle\Snippet\SnippetRepository;
-use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
+use Sulu\Component\Content\Document\WorkflowStage;
+use Sulu\Component\Content\Form\Exception\InvalidFormException;
 use Sulu\Component\Content\Mapper\ContentMapper;
-use Sulu\Component\Content\Mapper\ContentMapperRequest;
+use Sulu\Component\DocumentManager\DocumentManager;
 use Sulu\Component\Rest\Exception\RestException;
 use Sulu\Component\Rest\ListBuilder\ListRepresentation;
 use Sulu\Component\Rest\ListBuilder\ListRestHelper;
 use Sulu\Component\Rest\RequestParametersTrait;
 use Sulu\Component\Security\SecuredControllerInterface;
+use Symfony\Component\Form\FormFactory;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -36,49 +40,54 @@ use Symfony\Component\Security\Core\SecurityContext;
 /**
  * handles snippets.
  */
-class SnippetController implements SecuredControllerInterface
+class SnippetController implements SecuredControllerInterface, ClassResourceInterface
 {
     use RequestParametersTrait;
 
     /**
      * @var ContentMapper
      */
-    protected $contentMapper;
+    private $contentMapper;
 
     /**
      * @var StructureManagerInterface
      */
-    protected $structureManager;
+    private $structureManager;
 
     /**
      * @var ViewHandler
      */
-    protected $viewHandler;
+    private $viewHandler;
 
     /**
      * @Var SnippetRepository
      */
-    protected $snippetRepository;
+    private $snippetRepository;
 
     /**
      * @var SecurityContext
      */
-    protected $securityContext;
+    private $securityContext;
 
     /**
      * @var UrlGeneratorInterface
      */
-    protected $urlGenerator;
-
-    /**
-     * @var string
-     */
-    protected $languageCode;
+    private $urlGenerator;
 
     /**
      * @var DefaultSnippetManagerInterface
      */
-    protected $defaultSnippetManager;
+    private $defaultSnippetManager;
+
+    /**
+     * @var DocumentManager
+     */
+    private $documentManager;
+
+    /**
+     * @var FormFactory
+     */
+    private $formFactory;
 
     public function __construct(
         ViewHandler $viewHandler,
@@ -87,7 +96,9 @@ class SnippetController implements SecuredControllerInterface
         SnippetRepository $snippetRepository,
         SecurityContext $securityContext,
         UrlGeneratorInterface $urlGenerator,
-        DefaultSnippetManagerInterface $defaultSnippetManager
+        DefaultSnippetManagerInterface $defaultSnippetManager,
+        DocumentManager $documentManager,
+        FormFactory $formFactory
     ) {
         $this->viewHandler = $viewHandler;
         $this->contentMapper = $contentMapper;
@@ -96,6 +107,8 @@ class SnippetController implements SecuredControllerInterface
         $this->securityContext = $securityContext;
         $this->urlGenerator = $urlGenerator;
         $this->defaultSnippetManager = $defaultSnippetManager;
+        $this->documentManager = $documentManager;
+        $this->formFactory = $formFactory;
     }
 
     /**
@@ -105,10 +118,9 @@ class SnippetController implements SecuredControllerInterface
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function getSnippetsAction(Request $request)
+    public function cgetAction(Request $request)
     {
-        $this->initEnv($request);
-
+        $locale = $this->getLocale($request);
         $listRestHelper = new ListRestHelper($request);
 
         // if the type parameter is falsy, assign NULL to $type
@@ -118,11 +130,11 @@ class SnippetController implements SecuredControllerInterface
 
         if ($uuidsString) {
             $uuids = explode(',', $uuidsString);
-            $snippets = $this->snippetRepository->getSnippetsByUuids($uuids, $this->languageCode);
+            $snippets = $this->snippetRepository->getSnippetsByUuids($uuids, $locale);
             $total = count($snippets);
         } else {
             $snippets = $this->snippetRepository->getSnippets(
-                $this->languageCode,
+                $locale,
                 $type,
                 $listRestHelper->getOffset(),
                 $listRestHelper->getLimit(),
@@ -132,7 +144,7 @@ class SnippetController implements SecuredControllerInterface
             );
 
             $total = $this->snippetRepository->getSnippetsAmount(
-                $this->languageCode,
+                $locale,
                 $type,
                 $listRestHelper->getSearchPattern(),
                 $listRestHelper->getSortColumn(),
@@ -140,16 +152,8 @@ class SnippetController implements SecuredControllerInterface
             );
         }
 
-        $data = [];
-
-        foreach ($snippets as $snippet) {
-            $snippetData = $snippet->toArray();
-            $snippetData['localizedTemplate'] = $snippet->getLocalizedTitle($this->languageCode);
-            $data[] = $snippetData;
-        }
-
         $data = new ListRepresentation(
-            $this->decorateSnippets($data, $this->languageCode),
+            $snippets,
             'snippets',
             'get_snippets',
             $request->query->all(),
@@ -171,13 +175,12 @@ class SnippetController implements SecuredControllerInterface
      *
      * @Get(defaults={"uuid" = ""})
      */
-    public function getSnippetAction(Request $request, $uuid = null)
+    public function getAction(Request $request, $uuid = null)
     {
-        $this->initEnv($request);
+        $locale = $this->getLocale($request);
 
-        $snippet = $this->contentMapper->load($uuid, null, $this->languageCode);
-
-        $view = View::create($this->decorateSnippet($snippet->toArray(), $this->languageCode));
+        $snippet = $this->documentManager->find($uuid, $locale);
+        $view = View::create($snippet);
 
         return $this->viewHandler->handle($view);
     }
@@ -189,23 +192,12 @@ class SnippetController implements SecuredControllerInterface
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function postSnippetAction(Request $request)
+    public function postAction(Request $request)
     {
-        $this->initEnv($request);
-        $data = $request->request->all();
+        $document = $this->documentManager->create('snippet');
+        $form = $this->processForm($request, $document);
 
-        $mapperRequest = ContentMapperRequest::create()
-            ->setType('snippet')
-            ->setTemplateKey($this->getRequired($request, 'template'))
-            ->setLocale($this->languageCode)
-            ->setUserId($this->getUser()->getId())
-            ->setData($data)
-            ->setState(intval($request->get('state', StructureInterface::STATE_PUBLISHED)));
-
-        $snippet = $this->contentMapper->saveRequest($mapperRequest);
-        $view = View::create($this->decorateSnippet($snippet->toArray(), $this->languageCode));
-
-        return $this->viewHandler->handle($view);
+        return $this->handleView($form->getData());
     }
 
     /**
@@ -216,24 +208,12 @@ class SnippetController implements SecuredControllerInterface
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function putSnippetAction(Request $request, $uuid)
+    public function putAction(Request $request, $uuid)
     {
-        $this->initEnv($request);
-        $data = $request->request->all();
+        $document = $this->findDocument($uuid, $this->getLocale($request));
+        $this->processForm($request, $document);
 
-        $mapperRequest = ContentMapperRequest::create()
-            ->setType('snippet')
-            ->setTemplateKey($this->getRequired($request, 'template'))
-            ->setUuid($uuid)
-            ->setLocale($this->languageCode)
-            ->setUserId($this->getUser()->getId())
-            ->setData($data)
-            ->setState(intval($request->get('state', StructureInterface::STATE_PUBLISHED)));
-
-        $snippet = $this->contentMapper->saveRequest($mapperRequest);
-        $view = View::create($this->decorateSnippet($snippet->toArray(), $this->languageCode));
-
-        return $this->viewHandler->handle($view);
+        return $this->handleView($document);
     }
 
     /**
@@ -244,8 +224,9 @@ class SnippetController implements SecuredControllerInterface
      *
      * @return JsonResponse
      */
-    public function deleteSnippetAction(Request $request, $uuid)
+    public function deleteAction(Request $request, $uuid)
     {
+        $locale = $this->getLocale($request);
         $webspaceKey = $request->query->get('webspace', null);
 
         $references = $this->snippetRepository->getReferences($uuid);
@@ -255,7 +236,7 @@ class SnippetController implements SecuredControllerInterface
             if ($force) {
                 $this->contentMapper->delete($uuid, $webspaceKey, true);
             } else {
-                return $this->getReferentialIntegrityResponse($webspaceKey, $references, $uuid);
+                return $this->getReferentialIntegrityResponse($webspaceKey, $references, $uuid, $locale);
             }
         } else {
             $this->contentMapper->delete($uuid, $webspaceKey);
@@ -279,7 +260,7 @@ class SnippetController implements SecuredControllerInterface
         $view = null;
         $snippet = null;
 
-        $this->initEnv($request);
+        $locale = $this->getLocale($request);
         $action = $this->getRequestParameter($request, 'action', true);
 
         try {
@@ -291,7 +272,7 @@ class SnippetController implements SecuredControllerInterface
                     $snippet = $this->snippetRepository->copyLocale(
                         $uuid,
                         $this->getUser()->getId(),
-                        $this->languageCode,
+                        $locale,
                         explode(',', $destLocale)
                     );
                     break;
@@ -301,7 +282,7 @@ class SnippetController implements SecuredControllerInterface
 
             // prepare view
             $view = View::create(
-                $this->decorateSnippet($snippet->toArray(), $this->languageCode),
+                $this->decorateSnippet($snippet->toArray(), $locale),
                 $snippet !== null ? 200 : 204
             );
         } catch (RestException $exc) {
@@ -316,7 +297,7 @@ class SnippetController implements SecuredControllerInterface
      *
      * @return JsonResponse
      */
-    public function getSnippetFieldsAction()
+    public function getFieldsAction()
     {
         return new JsonResponse(
             [
@@ -394,50 +375,6 @@ class SnippetController implements SecuredControllerInterface
    }
 
     /**
-     * Initiates the environment.
-     */
-    private function initEnv(Request $request)
-    {
-        $this->languageCode = $this->getLocale($request);
-
-        if (!$this->languageCode) {
-            throw new \InvalidArgumentException('You must provide the "language" query parameter');
-        }
-    }
-
-    /**
-     * Returns a required parameter.
-     */
-    private function getRequired(Request $request, $parameterName)
-    {
-        $value = $request->request->get($parameterName);
-
-        if (null === $value) {
-            throw new \InvalidArgumentException(
-                sprintf(
-                    'You must provide a value for the POST parameter "%s"',
-                    $parameterName
-                )
-            );
-        }
-
-        return $value;
-    }
-
-    /**
-     * Decorate snippets for HATEOAS.
-     */
-    private function decorateSnippets(array $snippets, $locale)
-    {
-        $res = [];
-        foreach ($snippets as $snippet) {
-            $res[] = $this->decorateSnippet($snippet, $locale);
-        }
-
-        return $res;
-    }
-
-    /**
      * Decorate snippet for HATEOAS.
      */
     private function decorateSnippet(array $snippet, $locale)
@@ -465,43 +402,6 @@ class SnippetController implements SecuredControllerInterface
     }
 
     /**
-     * Return a response for the case where there is an referential integrity violation.
-     *
-     * It will return a 409 (Conflict) response with an array of structures which reference
-     * the node and an array of "other" nodes (i.e. non-structures) which reference the node.
-     *
-     * @param string $webspace
-     * @param NodeInterface[] $references
-     * @param string $uuid
-     *
-     * @return Response
-     */
-    private function getReferentialIntegrityResponse($webspace, $references, $uuid)
-    {
-        $data = [
-            'structures' => [],
-            'other' => [],
-            'isDefault' => $this->defaultSnippetManager->isDefault($uuid),
-        ];
-
-        foreach ($references as $reference) {
-            if ($reference->getParent()->isNodeType('sulu:page')) {
-                $content = $this->contentMapper->load(
-                    $reference->getParent()->getIdentifier(),
-                    $webspace,
-                    $this->languageCode,
-                    true
-                );
-                $data['structures'][] = $content->toArray();
-            } else {
-                $data['other'] = $reference->getPath();
-            }
-        }
-
-        return new JsonResponse($data, 409);
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function getLocale(Request $request)
@@ -515,5 +415,91 @@ class SnippetController implements SecuredControllerInterface
     public function getSecurityContext()
     {
         return 'sulu.global.snippets';
+    }
+
+    /**
+     * Return a response for the case where there is an referential integrity violation.
+     *
+     * It will return a 409 (Conflict) response with an array of structures which reference
+     * the node and an array of "other" nodes (i.e. non-structures) which reference the node.
+     *
+     * @param string $webspace
+     * @param NodeInterface[] $references
+     * @param string $uuid
+     *
+     * @return Response
+     */
+    private function getReferentialIntegrityResponse($webspace, $references, $uuid, $locale)
+    {
+        $data = [
+            'structures' => [],
+            'other' => [],
+            'isDefault' => $this->defaultSnippetManager->isDefault($uuid),
+        ];
+
+        foreach ($references as $reference) {
+            if ($reference->getParent()->isNodeType('sulu:page')) {
+                $content = $this->contentMapper->load(
+                    $reference->getParent()->getIdentifier(),
+                    $webspace,
+                    $locale,
+                    true
+                );
+                $data['structures'][] = $content->toArray();
+            } else {
+                $data['other'] = $reference->getPath();
+            }
+        }
+
+        return new JsonResponse($data, 409);
+    }
+
+    private function findDocument($uuid, $locale)
+    {
+        return $this->documentManager->find(
+            $uuid,
+            $locale,
+            [
+                'load_ghost_content' => false,
+            ]
+        );
+    }
+
+    private function processForm(Request $request, $document)
+    {
+        $locale = $this->getLocale($request);
+        $data = $request->request->all();
+        $data['workflowStage'] = $request->get('state', WorkflowStage::PUBLISHED);
+
+        $form = $this->formFactory->create('snippet', $document, [
+            'csrf_protection' => false,
+        ]);
+        $form->submit($data, false);
+
+        if (!$form->isValid()) {
+            throw new InvalidFormException($form);
+        }
+
+        $this->documentManager->persist(
+            $document,
+            $locale,
+            [
+                'user' => $this->getUser()->getId(),
+                'clear_missing_content' => false,
+            ]
+        );
+        $this->documentManager->flush();
+
+        return $form;
+    }
+
+    private function handleView($document)
+    {
+        $view = View::create($document);
+        $view->setSerializationContext(
+            SerializationContext::create()->setSerializeNull(true)
+        );
+
+        return $this->viewHandler->handle($view);
     }
 }