+++ src/Sulu/Bundle/SnippetBundle/Tests/Functional/Controller/SnippetControllerTest.php
@@ -48,16 +48,44 @@ class SnippetControllerTest extends SuluTestCase
         $this->client = $this->createAuthenticatedClient();
     }
 
-    public function testGet()
+    /**
+     * @dataProvider provideGet
+     */
+    public function testGet($locale, $expected)
     {
-        $this->client->request('GET', '/snippets/' . $this->hotel1->getUuid() . '?language=de');
+        $this->client->request('GET', '/snippets/' . $this->hotel1->getUuid() . '?language=' . $locale);
         $response = $this->client->getResponse();
 
-        $this->assertEquals(200, $response->getStatusCode());
+        $result = $response->getContent();
         $result = json_decode($response->getContent(), true);
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertLinks([
+            'self', 'delete', 'update', 'new',
+        ], $result);
 
-        $this->assertEquals('Le grande budapest', $result['title']); // snippet nodes do not have a path
+        $this->assertEquals($expected['title'], $result['title']);
+        $this->assertEquals($expected['description'], $result['description']);
         $this->assertEquals($this->hotel1->getUuid(), $result['id']);
+    }
+
+    public function provideGet()
+    {
+        return [
+            [
+                'de',
+                [
+                    'title' => 'Das Großes Budapest',
+                    'description' => 'Hallo Weld!',
+                ],
+            ],
+            [
+                'en',
+                [
+                    'title' => 'The Grand Budapest',
+                    'description' => 'Hello World',
+                ],
+            ],
+        ];
     }
 
     public function testGetMany()
@@ -70,12 +98,16 @@ class SnippetControllerTest extends SuluTestCase
         ));
         $response = $this->client->getResponse();
 
-        $this->assertEquals(200, $response->getStatusCode());
         $result = json_decode($response->getContent(), true);
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertLinks([
+            'self', 'first', 'last', 'filter', 'find', 'pagination', 'sortable',
+        ], $result);
+
         $this->assertCount(2, $result['_embedded']['snippets']);
 
         $result = reset($result['_embedded']['snippets']);
-        $this->assertEquals('Le grande budapest', $result['title']);
+        $this->assertEquals('Das Großes Budapest', $result['title']);
         $this->assertEquals($this->hotel1->getUuid(), $result['id']);
         $this->assertEquals('Hotel', $result['localizedTemplate']);
     }
@@ -114,7 +146,7 @@ class SnippetControllerTest extends SuluTestCase
         $result = $result['_embedded']['snippets'];
         $this->assertCount(2, $result);
         $result = reset($result);
-        $this->assertEquals('Le grande budapest', $result['title']);
+        $this->assertEquals('Das Großes Budapest', $result['title']);
         $this->assertEquals($this->hotel1->getUuid(), $result['id']);
     }
 
@@ -175,6 +207,7 @@ class SnippetControllerTest extends SuluTestCase
         $response = $this->client->getResponse();
 
         $this->assertEquals(200, $response->getStatusCode());
+
         $result = json_decode($response->getContent(), true);
         $this->assertCount($expectedNbResults, $result['_embedded']['snippets']);
 
@@ -218,17 +251,29 @@ class SnippetControllerTest extends SuluTestCase
         $data = [
             'template' => 'car',
             'title' => 'My New Car',
+            'description' => 'My car is red.',
         ];
 
         $query = http_build_query($params);
         $this->client->request('POST', '/snippets?' . $query, $data);
         $response = $this->client->getResponse();
 
-        $this->assertEquals(200, $response->getStatusCode());
         $result = json_decode($response->getContent(), true);
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertLinks([
+            'self', 'delete', 'update', 'new',
+        ], $result);
+
         $this->assertEquals($data['title'], $result['title']);
         $this->assertEquals($params['language'], reset($result['concreteLanguages']));
         $this->assertEquals(StructureInterface::STATE_PUBLISHED, $result['nodeState']);
+        $this->assertEquals('My car is red.', $result['description']);
+
+        try {
+            $this->documentManager->find($result['id'], 'de');
+        } catch (\DocumentNotFoundException $e) {
+            $this->fail('Document was not persisted');
+        }
     }
 
     /**
@@ -283,49 +328,45 @@ class SnippetControllerTest extends SuluTestCase
         $this->assertEquals(StructureInterface::STATE_TEST, $result['nodeState']);
     }
 
-    public function providePut()
+    public function testPut()
     {
-        return [
-            [
-                [],
-                [
-                    'template' => 'hotel',
-                    'title' => 'Renamed Hotel',
-                ],
-            ],
+        $data = [
+            'template' => 'hotel',
+            'title' => 'Renamed Hotel',
+            'description' => 'My hotel is red',
         ];
-    }
 
-    /**
-     * @dataProvider providePut
-     */
-    public function testPut($params, $data)
-    {
-        $params = array_merge([
+        $params = [
             'language' => 'de',
-        ], $params);
+        ];
 
         $query = http_build_query($params);
         $this->client->request('PUT', sprintf('/snippets/%s?%s', $this->hotel1->getUuid(), $query), $data);
         $response = $this->client->getResponse();
 
-        $this->assertEquals(200, $response->getStatusCode());
         $result = json_decode($response->getContent(), true);
-        $this->assertEquals($data['template'], $result['template']);
-        $this->assertEquals($data['title'], $result['title']);
-        $this->assertEquals($params['language'], reset($result['concreteLanguages']));
+        $this->assertEquals(200, $response->getStatusCode());
+
+        foreach ($data as $key => $value) {
+            $this->assertEquals($data[$key], $value);
+        }
+
+        $this->assertContains($params['language'], $result['concreteLanguages']);
         $this->assertEquals(StructureInterface::STATE_PUBLISHED, $result['nodeState']);
+
+        $document = $this->documentManager->find($result['id'], 'de');
+
+        $this->assertEquals($data['title'], $document->getTitle());
+        $this->assertEquals($data['description'], $document->getStructure()->getProperty('description')->getValue());
     }
 
-    /**
-     * @dataProvider providePut
-     */
-    public function testPutPublished($params, $data)
+    public function testPutPublished()
     {
-        $params = array_merge([
+        $data = [];
+        $params = [
             'language' => 'de',
             'state' => StructureInterface::STATE_PUBLISHED,
-        ], $params);
+        ];
 
         $query = http_build_query($params);
         $this->client->request('PUT', sprintf('/snippets/%s?%s', $this->hotel1->getUuid(), $query), $data);
@@ -333,21 +374,25 @@ class SnippetControllerTest extends SuluTestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $result = json_decode($response->getContent(), true);
-        $this->assertEquals($data['template'], $result['template']);
-        $this->assertEquals($data['title'], $result['title']);
-        $this->assertEquals($params['language'], reset($result['concreteLanguages']));
+        $this->assertLinks([
+            'self', 'delete', 'update', 'new',
+        ], $result);
+
+        foreach ($data as $key => $value) {
+            $this->assertEquals($value, $result[$key]);
+        }
+
+        $this->assertContains($params['language'], $result['concreteLanguages']);
         $this->assertEquals(StructureInterface::STATE_PUBLISHED, $result['nodeState']);
     }
 
-    /**
-     * @dataProvider providePut
-     */
-    public function testPutTest($params, $data)
+    public function testPutTest()
     {
-        $params = array_merge([
+        $data = [];
+        $params = [
             'language' => 'de',
             'state' => StructureInterface::STATE_TEST,
-        ], $params);
+        ];
 
         $query = http_build_query($params);
         $this->client->request('PUT', sprintf('/snippets/%s?%s', $this->hotel1->getUuid(), $query), $data);
@@ -355,9 +400,6 @@ class SnippetControllerTest extends SuluTestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $result = json_decode($response->getContent(), true);
-        $this->assertEquals($data['template'], $result['template']);
-        $this->assertEquals($data['title'], $result['title']);
-        $this->assertEquals($params['language'], reset($result['concreteLanguages']));
         $this->assertEquals(StructureInterface::STATE_TEST, $result['nodeState']);
     }
 
@@ -395,12 +437,27 @@ class SnippetControllerTest extends SuluTestCase
         $this->assertEquals('Hotel de', $newPage->getTitle());
     }
 
+    public function testGetFields()
+    {
+        $this->client->request('GET', '/snippet/fields');
+        $response = $this->client->getResponse();
+        $this->assertEquals(200, $response->getStatusCode());
+        $body = $response->getContent();
+        $fields = json_decode($body);
+        $this->assertNotNull($fields);
+    }
+
     private function loadFixtures()
     {
         // HOTELS
         $this->hotel1 = $this->documentManager->create('snippet');
         $this->hotel1->setStructureType('hotel');
-        $this->hotel1->setTitle('Le grande budapest');
+        $this->hotel1->setTitle('The Grand Budapest');
+        $this->hotel1->getStructure()->getProperty('description')->setValue('Hello World');
+        $this->documentManager->persist($this->hotel1, 'en');
+
+        $this->hotel1->getStructure()->getProperty('description')->setValue('Hallo Weld!');
+        $this->hotel1->setTitle('Das Großes Budapest');
         $this->documentManager->persist($this->hotel1, 'de');
 
         $this->hotel2 = $this->documentManager->create('snippet');
@@ -430,5 +487,20 @@ class SnippetControllerTest extends SuluTestCase
         $this->documentManager->persist($car, 'en');
 
         $this->documentManager->flush();
+    }
+
+    private function assertLinks(array $links, array $response)
+    {
+        $this->assertArrayHasKey('_links', $response);
+        $responseLinks = $response['_links'];
+
+        $diff = array_diff($links, array_keys($responseLinks));
+
+        if ($diff) {
+            $this->fail(sprintf(
+                'Failed asserting that the links "%s" were present in response. Got "%s"',
+                implode('", "', $diff), implode('", "', array_keys($responseLinks))
+            ));
+        }
     }
 }